==> r_scripts/2_get_hybrids_floras.R <==
# Marina Golivets
# BiCIKL Hackathon (Sept 20-24, 2021, Meise, Belgium)
# Extract hybrid and parent taxon names from various floras

# Note: Here we reuse input data sets generated by M. Golivets within a different project

# get hybrid names from FlorKart, German SL, Pladias, Swiss Flora ----------------------------------
hybr_florkart <- read_csv2(here("input/taxa_florkart.csv")) %>%
  filter(
    isHybrid == TRUE,
    taxonRank == "species"
  ) %>%
  pull(taxonName, acceptedTaxonName) %>%
  tibble(
    verbatim = .,
    source = "florkart"
    )

hybr_gsl <- readxl::read_xlsx(here("input/GermanSL1.5_replaceID.xlsx")) %>%
  janitor::clean_names() %>%
  filter(
    str_detect(taxon_name, "×|^x |^X "),
    taxon_rank == "SPE"
  ) %>%
  mutate(
    name_author12 = gsub("-", "", name_author12),
    taxon_name = str_trim(paste(taxon_name, name_author12))
  ) %>%
  pull(taxon_name, neuer_name) %>%
  tibble(
    verbatim = .,
    source = "germansl"
  )

hybr_pladias <- read_csv2(here("input/taxa_pladias.csv")) %>%
  filter(
    isHybrid == TRUE,
    taxonRank == "species"
  ) %>%
  pull(taxonName, acceptedTaxonName) %>%
  tibble(
    verbatim = .,
    source = "pladias"
  )

hybr_swissflora <- readxl::read_xlsx(
  here("input/Checklist_2017_simple_version_20181017.xlsx"),
  skip = 4
) %>%
  slice(2:n()) %>%
  filter(str_detect(Taxonname, "×|^x |^X ")) %>%
  transmute(
    verbatim = Taxonname,
    source = "swissflora"
  )

# combine and edit
hybr_floras <- bind_rows(
  hybr_florkart, hybr_gsl, hybr_pladias, hybr_swissflora
) %>%
  mutate(edited = str_replace_all(
    verbatim, c(
      "\\(purpurea × repens\\)" = "purpurea × repens",
      "tictoria" = "tinctoria",
      " riviana" = " riviniana",
      " spathulus" = " spathulatus",
      " erracticus" = " erraticus"
      ))) %>%
  filter(str_detect(edited, "'| hort\\.|Hort\\.", negate = TRUE)) %>%
  mutate(
    edited = str_remove_all(edited, "\\?") %>%
      str_replace_all(" x |×|^X ", " × ") %>%
      str_squish() %>%
      str_trim() %>%
      str_remove(" NA$|, excl\\. descr\\.") %>%
      str_replace(" × [A-Z]\\. | × [A-Z][a-z]\\. ", " × "),
    type = if_else(str_detect(edited, "^[A-Za-z]+ × |^× |, pro "), "name", "formula"),
    formula = edited %>%
      if_else(str_detect(., "^[A-Za-z]+ × |^× "), NA_character_, .) %>%
      if_else(str_detect(edited, " pro "), str_extract(edited, ", pro .*"), .) %>%
      str_remove(", pro "),
    edited = str_remove(edited, ", pro .*"),
    parent = str_split(formula, pattern = "×")
  ) %>%
  unnest(cols = "parent") %>%
  mutate(
    genus = str_extract(edited,  "\\w*"),
    parent = parent %>%
      str_squish() %>% 
      str_trim(),
    parent = case_when(
    grepl("^[a-z]", parent) ~ paste(genus, parent),
    grepl("^[A-Z]\\.", parent) ~ paste(genus, gsub("^[A-Z]\\.", "", parent)),
    TRUE ~ parent
  )) %>%
  mutate(
    parent = if_else(
      str_detect(verbatim, " x |× | X |, pro", negate = TRUE), NA_character_, parent) %>%
      if_else(str_detect(., "× "), NA_character_, .) %>%
      if_else(str_replace_all(edited, " × ", " ") == ., NA_character_, .) %>%
      str_squish() %>% 
      str_trim(),
    edited2 = if_else(
      type == "name", str_replace_all(edited, c(" × " = " ×", "^× " = "×")), edited) %>%
      str_remove(", s\\. str\\.|, s\\. l\\."),
    type = if_else(is.na(parent), "name", type)
  ) %>%
  filter(str_detect(edited, "× .* ×", negate = TRUE)) %>%
  select(-genus, -formula) %>%
  filter(str_count(parent, pattern = " ") > 0 | is.na(parent)) %>%
  filter(
    edited2 != "Viola ×wittrockiana Viola ×wittrockiana GAMS",
    str_detect(edited, " subsp\\. ", negate = TRUE)) %>% # exclude hybrids involving subspecies
  distinct()

# match names to GBIF ------------------------------------------------------------------------------
gbif_h <- gbif.cmpfn(
  unique(hybr_floras$edited2),
  seq_along(unique(hybr_floras$edited2))
) %>%
  filter(str_detect(scientificName, "×| x ") | confidence > 95 | confidence == 0) %>%
  group_by(name) %>%
  filter(n() == 1) %>% 
  ungroup()
setdiff(unique(hybr_floras$edited2), gbif_h$name)

hybr_floras %<>% 
  left_join(
    gbif_h %>% 
      transmute(
        name, usageKey,
        acceptedUsageKey = coalesce(acceptedUsageKey, usageKey)),
    by = c("edited2" = "name")) 

gbif_p <- gbif.cmpfn(
  unique(hybr_floras$parent),
  seq_along(unique(hybr_floras$parent))
) %>%
  filter(str_detect(scientificName, "×| x ") | confidence > 95 | confidence == 0) %>%
  group_by(name) %>%
  filter(n() == 1) %>% 
  ungroup()
setdiff(unique(hybr_floras$parent), gbif_p$name)

hybr_floras %<>% 
  group_by(edited) %>%
  filter(n() == 2) %>%
  mutate(parent_type = c("parent1", "parent2")) %>%
  ungroup() %>%
  pivot_wider(values_from = parent, names_from = parent_type)

hybr_floras %<>% 
  left_join(
    gbif_p %>% 
      transmute(
        parent1 = name, 
        parent1_usageKey = usageKey,
        parent1_acceptedUsageKey = coalesce(acceptedUsageKey, usageKey))
  ) %>%
  left_join(
    gbif_p %>% 
      transmute(
        parent2 = name, 
        parent2_usageKey = usageKey,
        parent2_acceptedUsageKey = coalesce(acceptedUsageKey, usageKey))
  ) 

write_csv2(hybr_floras, here("output/hybrids_floras.csv"))
